Nali on linux, readme, formula renames upd
