+++ formula_renames.json
@@ -1,0 +1,4 @@
+{
+  "adguardhome-bin": "adguardhome",
+  "nali-bin": "nali"
+}
